+++ examples/react/release/boxes/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@agile-ts/core": "^0.1.2",
+    "@agile-ts/core": "file:.yalc/@agile-ts/core",
     "@agile-ts/logger": "^0.0.7",
     "@agile-ts/proxytree": "^0.0.5",
     "@agile-ts/react": "^0.1.2",
