+++ packages/core/src/index.ts
@@ -1,4 +1,14 @@
-import { Agile } from './internal';
+import { Agile } from './agile';
 
-export * from './internal';
+export * from './storages';
+export * from './storages/storage';
+export * from './state';
+export * from './integrations/integration';
+export * from './computed';
+export * from './collection';
+export * from './collection/item';
+export * from './collection/group';
+export * from './collection/selector';
+export * from './shared';
+
 export default Agile;
