+++ packages/core/src/state/state.runtime.job.ts
@@ -1,10 +1,10 @@
 import {
-  defineConfig,
   RuntimeJob,
   RuntimeJobConfigInterface,
   RuntimeJobKey,
-  StateObserver,
-} from '../internal';
+} from '../runtime/runtime.job';
+import { StateObserver } from './state.observer';
+import { defineConfig } from '@agile-ts/utils';
 
 export class StateRuntimeJob extends RuntimeJob<StateObserver> {
   public config: StateRuntimeJobConfigInterface;
