+++ packages/core/src/storages/persistent.ts
@@ -1,10 +1,7 @@
-import {
-  Agile,
-  copy,
-  defineConfig,
-  LogCodeManager,
-  StorageKey,
-} from '../internal';
+import { Agile } from '../agile';
+import { StorageKey } from './storage';
+import { copy, defineConfig } from '@agile-ts/utils';
+import { LogCodeManager } from '../logCodeManager';
 
 export class Persistent {
   // Agile Instance the Persistent belongs to
