+++ packages/core/src/storages/storage.ts
@@ -2,9 +2,9 @@ import {
   isJsonString,
   isAsyncFunction,
   isFunction,
-  LogCodeManager,
   defineConfig,
-} from '../internal';
+} from '@agile-ts/utils';
+import { LogCodeManager } from '../logCodeManager';
 
 export class Storage {
   public config: StorageConfigInterface;
@@ -182,6 +182,25 @@ export class Storage {
       ? `_${this.config.prefix}_${key}`
       : key.toString();
   }
+}
+
+/**
+ * Returns a newly created Storage.
+ *
+ * A Storage Class serves as an interface to external storages,
+ * such as the [Async Storage](https://github.com/react-native-async-storage/async-storage) or
+ * [Local Storage](https://www.w3schools.com/html/html5_webstorage.asp).
+ *
+ * It creates the foundation to easily [`persist()`](https://agile-ts.org/docs/core/state/methods#persist) [Agile Sub Instances](https://agile-ts.org/docs/introduction/#agile-sub-instance)
+ * (like States or Collections) in nearly any external storage.
+ *
+ * [Learn more..](https://agile-ts.org/docs/core/agile-instance/methods#createstorage)
+ *
+ * @public
+ * @param config - Configuration object
+ */
+export function createStorage(config: CreateStorageConfigInterface): Storage {
+  return new Storage(config);
 }
 
 export type StorageKey = string | number;
